(NPUP-15) Port divide operator to the dispatcher.
Porting the divide operator's implementation to the dispatcher.
