+++ lib/include/puppet/compiler/evaluation/operators/binary/divide.hpp
@@ -1,0 +1,23 @@
+/**
+ * @file
+ * Declares the divide operator.
+ */
+#pragma once
+
+#include "descriptor.hpp"
+
+namespace puppet { namespace compiler { namespace evaluation { namespace operators { namespace binary {
+
+    /**
+     * Implements the divide operator.
+     */
+    struct divide
+    {
+        /**
+         * Create a binary operator descriptor.
+         * @return Returns the binary operator descriptor representing this operator.
+         */
+        static descriptor create_descriptor();
+    };
+
+}}}}}  // namespace puppet::compiler::evaluation::operators::binary
