--- lib/include/puppet/compiler/evaluation/operators/divide.hpp
@@ -1,24 +1,0 @@
-/**
- * @file
- * Declares the divide operator.
- */
-#pragma once
-
-#include "binary_operator_context.hpp"
-
-namespace puppet { namespace compiler { namespace evaluation { namespace operators {
-
-    /**
-     * Implements the divide operator.
-     */
-    struct divide
-    {
-        /**
-         * Called to invoke the operator.
-         * @param context The binary operator context.
-         * @return Returns the resulting value.
-         */
-        runtime::values::value operator()(binary_operator_context const& context) const;
-    };
-
-}}}}  // namespace puppet::compiler::evaluation::operators
