+++ lib/src/compiler/evaluation/dispatcher.cc
@@ -27,6 +27,7 @@
 #include <puppet/compiler/evaluation/functions/with.hpp>
 #include <puppet/compiler/evaluation/operators/binary/call_context.hpp>
 #include <puppet/compiler/evaluation/operators/binary/assignment.hpp>
+#include <puppet/compiler/evaluation/operators/binary/divide.hpp>
 #include <puppet/compiler/evaluation/operators/unary/call_context.hpp>
 #include <puppet/compiler/exceptions.hpp>
 #include <boost/format.hpp>
@@ -68,6 +69,7 @@ namespace puppet { namespace compiler { namespace evaluation {
 
         // Add the built-in binary operators
         add(binary::assignment::create_descriptor());
+        add(binary::divide::create_descriptor());
     }
 
     void dispatcher::add(functions::descriptor descriptor)
