+++ lib/src/compiler/evaluation/evaluator.cc
@@ -6,7 +6,6 @@
 #include <puppet/compiler/evaluation/functions/call_context.hpp>
 #include <puppet/compiler/evaluation/operators/binary/call_context.hpp>
 #include <puppet/compiler/evaluation/operators/unary/call_context.hpp>
-#include <puppet/compiler/evaluation/operators/divide.hpp>
 #include <puppet/compiler/evaluation/operators/equals.hpp>
 #include <puppet/compiler/evaluation/operators/greater.hpp>
 #include <puppet/compiler/evaluation/operators/greater_equal.hpp>
@@ -919,7 +918,6 @@ namespace puppet { namespace compiler { namespace evaluation {
         ast::binary_operation const& operation)
     {
         static const unordered_map<binary_operator, function<value(operators::binary_operator_context const&)>, boost::hash<binary_operator>> binary_operators = {
-            { ast::binary_operator::divide,             operators::divide() },
             { ast::binary_operator::equals,             operators::equals() },
             { ast::binary_operator::greater_than,       operators::greater() },
             { ast::binary_operator::greater_equals,     operators::greater_equal() },
